+++ CHANGES.md
@@ -9,7 +9,8 @@ Changes since 1.4
   - Add ability to return matrix factorization and reuse it on subsequent
     calls to `mplinsolve()` when using `LU` as solver, by passing the returned
     `info` struct containing the factors to subsequent calls via the
-    `lu_factors` option.
+    `lu_factors` option. Transposed systems can be solved with the same
+    factors by setting `opt.lu_factors.t` to a true value.
 
 
 Version 1.4 - *Oct 8, 2020*